backend/fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
